intlen base and putnbr base now using unsigned long long int

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: ocgraf <ocgraf@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 10:09:54 by ocgraf            #+#    #+#             */
-/*   Updated: 2025/05/28 11:21:31 by ocgraf           ###   ########.fr       */
+/*   Updated: 2025/06/02 15:02:10 by ocgraf           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int		ft_isascii(int c);
 int		ft_isdigit(int c);
 int		ft_isprint(int c);
 size_t	ft_intlen(int nb);
-size_t	ft_intlen_base(int nb, char *base);
+size_t	ft_intlen_base(unsigned long long int nb, char *base);
 char	*ft_itoa(int n);
 void	*ft_memchr(const void *s, int c, size_t n);
 int		ft_memcmp(const void *s1, const void *s2, size_t n);
@@ -36,7 +36,7 @@ void	ft_putchar(char c);
 void	ft_putchar_fd(char c, int fd);
 void	ft_putendl_fd(char *s, int fd);
 void	ft_putnbr(int nb);
-void	ft_putnbr_base(int nbr, char *base);
+void	ft_putnbr_base(unsigned long long int nbr, char *base);
 void	ft_putnbr_fd(int n, int fd);
 void	ft_putstr(char *s);
 void	ft_putstr_fd(char *s, int fd);
